Extract logger config to separate file
