+++ logger.py
@@ -1,0 +1,24 @@
+import os
+import logging
+from pythonjsonlogger import jsonlogger
+
+
+def configure_log():
+    """Configure JSON logger."""
+    log_level = os.environ.get("EMR_LAUNCHER_LOG_LEVEL", "INFO").upper()
+    numeric_level = getattr(logging, log_level, None)
+    if not isinstance(numeric_level, int):
+        raise ValueError("Invalid log level: %s" % log_level)
+    if len(logging.getLogger().handlers) > 0:
+        logging.getLogger().setLevel(log_level)
+    else:
+        logging.basicConfig(level=log_level)
+    logger = logging.getLogger()
+    logger.propagate = False
+    console_handler = logging.StreamHandler()
+    formatter = jsonlogger.JsonFormatter(
+        "%(asctime)s %(name)-12s %(levelname)-8s %(message)s"
+    )
+    console_handler.setFormatter(formatter)
+    logger.addHandler(console_handler)
+    return logger
